# webpos-api/src/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Product and Item
  license:
    name: MIT

servers:
  - url: http://localhost:8080/api

paths:
  /products:
    get:
      tags:
        - product
      operationId: listProducts
      summary: Lists products
      description: Returns an array of products.
      responses:
        200:
          description: Products found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - product
      operationId: addProduct
      summary: Create a Product
      description: Creates a product .
      requestBody:
        description: The product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        200:
          description: Product created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{productId}:
    get:
      tags:
        - product
      operationId: getProduct
      summary: Get a product by ID
      description: Returns the product or a 404 error.
      parameters:
        - name: productId
          in: path
          description: The ID of the product.
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: product details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

    put:
      tags:
        - product
      operationId: updateProduct
      summary: Update a product by ID
      description: Returns the product or a 404 error.
      parameters:
        - name: productId
          in: path
          description: The ID of the product.
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: The product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        200:
          description: Product details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

    delete:
      tags:
        - product
      operationId: deleteProduct
      summary: Delete a Product by ID
      description: Returns the Product or a 404 error.
      parameters:
        - name: productId
          in: path
          description: The ID of the product.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      responses:
        200:
          description: Product details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  /items:
    get:
      tags:
        - item
      operationId: listItems
      summary: List cart items
      description: Return an array of items.
      responses:
        200:
          description: Items found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      tags:
        - item
      operationId: AddItem
      summary: Add a cart item
      description: Add a item.
      requestBody:
        description: The item
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        200:
          description: Item created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
  /items/{itemId}:
    get:
      tags:
        - item
      operationId: getItem
      summary: Get a item by ID
      description: Returns the item or a 404 error.
      parameters:
        - name: itemId
          in: path
          description: The ID of the item.
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: item details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
    put:
      tags:
        - item
      operationId: updateItem
      summary: Update a item by ID
      description: Returns the item or a 404 error.
      parameters:
        - name: itemId
          in: path
          description: The ID of the item.
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: The item
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        200:
          description: Item details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
    delete:
      tags:
        - item
      operationId: deleteItem
      summary: Delete a Item by ID
      description: Returns the Item or a 404 error.
      parameters:
        - name: itemId
          in: path
          description: The ID of the item.
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Item details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'



components:
  schemas:
    Product:
      title: Product
      description: A product.
      type: object
      properties:
        id:
          title: ID
          description: The ID of the product.
          type: integer
          format: int32
        name:
          title: Name
          description: The name of the product.
          type: string
        price:
          title: Price
          description: The price of the product.
          type: double
        image:
          title: Image
          description: The image of the product.
          type: string
        quantity:
          title: Quantity
          description: The quantity of the product.
          type: integer
      required:
        - id
        - name
        - price
        - image
        - quantity
    Item:
      title: Item
      description: A Cart Item.
      type: object
      properties:
        id:
          title: ID
          description: the ID of the Item.
          type: integer
          format: int32
        pid:
          title: ProductId
          description: the product id of the Item.
          type: integer
        quantity:
          title: Quantity
          description: the quantity of the Item.
          type: integer
      required:
        - id
        - pid
        - quantity